+++ src/ui/popover/wasmenu-btn/wasmenu-btn.component.ts
@@ -1,19 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { WasReview } from '../../../ui/popover/wasreview/wasreview.dialog';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { UserService } from '../../../user.service';
 import { User } from '../../../app.models';
-import { WasStore } from '../../../ui/popover/wasstore/wasstore.dialog';
-import { WasSSO } from '../../../ui/popover/wassso/wassso.dialog';
-import { WasAlert } from '../../../ui/popover/wasalert/wasalert.dialog';
 
 @Component({
   selector: 'was-menu-btn',
   templateUrl: './wasmenu-btn.component.html',
   styleUrls: ['../was.component.css'],
 })
-export class WasMenuBtn implements OnInit {
+export class WasMenuBtn {
   public userloggedin = false;
   // public loginmessage = 'Login with SSO';
   /**
@@ -23,14 +18,8 @@ export class WasMenuBtn implements OnInit {
  * <was-menu-btn></was-menu-btn>
  *
 */
-  constructor(
-    public dialog: MatDialog,
-    private userService: UserService,
-  ) {
-  }
-  ngOnInit(): void {
-    // this.loginlogoutMessage();
-  }
+  constructor(private userService: UserService) {}
+
   get loginMessage() {
     return this.userService.isLoggedInObs.map((_isLogged: Boolean) => {
       if (_isLogged) {
@@ -44,27 +33,13 @@ export class WasMenuBtn implements OnInit {
   }
 
   leavereview(): void {
-    const thissso = this.dialog.open(WasReview);
+    this.userService.leavereview();
   }
   openstore() {
-    const thissso = this.dialog.open(WasStore);
+    this.userService.openstore();
   }
   opensso() {
-    if (this.userloggedin) {
-      const dialogRef = this.dialog.open(WasAlert, {
-        data: { title: 'Do you wish to log out?', body: 'Log out of your WickeyAppStore SSO account?', buttons: ['Yes', 'No'] }
-      });
-      dialogRef.afterClosed().subscribe(result => {
-        if (result === 0) {
-          // yes selected
-          // log out.
-          console.log('log out this user');
-          this.userService.logOut();
-        }
-      })
-    } else {
-      this.dialog.open(WasSSO);
-    }
+    this.userService.opensso();
   }
 
 }